modified the reaction type for id

+++ cvpWebApi/App_Data/DBConnection.cs
@@ -1260,7 +1260,7 @@ namespace cvp
                             while (dr.Read())
                             {
                                 var item = new Reaction();
-                                item.reaction_id = dr["REACTION_ID"] == DBNull.Value ? 0 : Convert.ToInt32(dr["REACTION_ID"]);
+                                item.reaction_id = dr["REACTION_ID"] == DBNull.Value ? 0 : Convert.ToInt64(dr["REACTION_ID"]);
                                 item.report_id = dr["REPORT_ID"] == DBNull.Value ? 0 : Convert.ToInt32(dr["REPORT_ID"]);
                                 item.duration = dr["DURATION"] == DBNull.Value ? 0 : Convert.ToInt32(dr["DURATION"]);
                                 item.duration_unit = dr["DURATION_UNIT"] == DBNull.Value ? string.Empty : dr["DURATION_UNIT"].ToString().Trim();
@@ -1289,7 +1289,7 @@ namespace cvp
         }
 
         // used by API
-        public Reaction GetReactionById(int id, string lang)
+        public Reaction GetReactionById(Int64 id, string lang)
         {
             var reactions = new Reaction();
             string commandText = "SELECT DISTINCT REACTION_ID, REPORT_ID, DURATION, MEDDRA_VERSION, ";
@@ -1321,7 +1321,7 @@ namespace cvp
                             while (dr.Read())
                             {
                                 var item = new Reaction();
-                                item.reaction_id = dr["REACTION_ID"] == DBNull.Value ? 0 : Convert.ToInt32(dr["REACTION_ID"]);
+                                item.reaction_id = dr["REACTION_ID"] == DBNull.Value ? 0 : Convert.ToInt64(dr["REACTION_ID"]);
                                 item.report_id = dr["REPORT_ID"] == DBNull.Value ? 0 : Convert.ToInt32(dr["REPORT_ID"]);
                                 item.duration = dr["DURATION"] == DBNull.Value ? 0 : Convert.ToInt32(dr["DURATION"]);
                                 item.duration_unit = dr["DURATION_UNIT"] == DBNull.Value ? string.Empty : dr["DURATION_UNIT"].ToString().Trim();
