+++ cvpWebApi/Models/IReactionRepository.cs
@@ -9,7 +9,7 @@ namespace cvpWebApi.Models
     interface IReactionRepository
     {
         IEnumerable<Reaction> GetAll(string lang);
-        Reaction Get(int id, string lang);
+        Reaction Get(Int64 id, string lang);
         IEnumerable<Reaction> GetReactionByReportId(string reportId, string lang);
     }
 }
